requirements.txt
pyautogui>=0.9.52
pydirectinput>=1.0.4
pygetwindow>=0.0.9
pywinauto>=0.6.8
opencv-python>=4.5.0
numpy>=1.21.0
psutil>=5.8.0
pyyaml>=6.0
ttkbootstrap>=1.0.0
PySide6>=6.5.0
pydantic>=1.10.0
typer>=0.2.1
pytest>=6.0.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0